+++ packages/simple_dart_code_gen/test/main_test.dart
@@ -40,6 +40,7 @@ import 'package:meta/meta.dart';
 @immutable
 final class SampleClass {
   /// document
+  @literal
   const SampleClass({
     required this.name,
     required this.age,