+++ packages/simple_dart_code_gen/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.5.2
 
+- output @literal
 - fix switch expr output
 - SimpleDartCode.toCodeString add format parameter for debug